+++ src/processor/linter/presets/validators/index.ts
@@ -1,4 +1,5 @@
-export { arrayOf, list, oneOrList, or, struct } from "./combinators";
+export { all, arrayOf, list, oneOrList, or, struct } from "./combinators";
 export { createValidator } from "./create";
+export { issue } from "./issue";
 export { boolean, color, number, string } from "./primitives";
 export { numberWithUnit } from "./unit";